[MOD/#193] 토글 스위치 공통 컴포넌트로 이동
